pyproject.toml
[project]
name = "sticky-notes"
version = "0.1.0"
description = "桌面便利贴应用"
requires-python = ">=3.10"
dependencies = [
    "customtkinter",
    "pystray",
    "pillow",
]
